batch-worker/requirements.txt
Django>=3.0,<4.0
celery~=5.4.0
kombu~=5.3.6
redis
psycopg2-binary
django-celery-results
django-celery-beat
dj-database-url
flower
python-logging-loki==0.3.1
python-dotenv==1.0.1
tzlocal~=5.2
pytz~=2024.1
requests~=2.32.3
python-dateutil~=2.9.0.post0
pyyaml
python-loki-logger
concurrent-log-handler